real-estate-ai/requirements.txt
requests 
bs4 
newspaper3k 
pdf2image
pytesseract